feat: PageNation 전체 화면 구성 완료 / Reactor 수정 완료
